fix circular refs for refactor
