+++ rabbit_config.py
@@ -4,7 +4,8 @@ import logging
 import cxoneflow_logging as cof_logging
 from config import ConfigurationException, get_config_path
 from config.server import CxOneFlowConfig
-from workflows.feedback_services import PRFeedbackService, PushFeedbackService, AbstractPRFeedbackService
+from workflows.feedback_services.push import PushFeedbackService
+from workflows.feedback_services.pr import PRQueueConstants
 from workflows.resolver_scan_service import ResolverScanService
 from workflows.scan_polling_service import ScanPollingService
 from workflows.base_service import CxOneFlowAbstractWorkflowService
@@ -35,7 +36,7 @@ async def setup() -> None:
 
             # Legacy are names/routes for backwards compatibility when upgrading.
             scan_in_exchange_legacy = await channel.declare_exchange(
-                AbstractPRFeedbackService.EXCHANGE_SCAN_INPUT_LEGACY,
+                PRQueueConstants.EXCHANGE_SCAN_INPUT_LEGACY,
                 aio_pika.ExchangeType.FANOUT,
                 durable=True,
             )
@@ -46,7 +47,7 @@ async def setup() -> None:
             )
             
             scan_await_exchange_legacy = await channel.declare_exchange(
-                AbstractPRFeedbackService.EXCHANGE_SCAN_WAIT_LEGACY,
+                PRQueueConstants.EXCHANGE_SCAN_WAIT_LEGACY,
                 aio_pika.ExchangeType.TOPIC,
                 durable=True,
                 internal=True,
@@ -60,13 +61,13 @@ async def setup() -> None:
 
 
             scan_annotate_exchange_pr = await channel.declare_exchange(
-                AbstractPRFeedbackService.EXCHANGE_SCAN_ANNOTATE,
+                PRQueueConstants.EXCHANGE_SCAN_ANNOTATE,
                 aio_pika.ExchangeType.TOPIC,
                 durable=True,
                 internal=True,
             )
             scan_feedback_exchange_pr = await channel.declare_exchange(
-                AbstractPRFeedbackService.EXCHANGE_SCAN_FEEDBACK,
+                PRQueueConstants.EXCHANGE_SCAN_FEEDBACK,
                 aio_pika.ExchangeType.TOPIC,
                 durable=True,
                 internal=True,
@@ -94,7 +95,7 @@ async def setup() -> None:
             # The awaited scans allows scans to soak until a timeout, then they go to the polling exchange where the
             # scan is polled to see the state or times out.
             polling_delivery_exchange_legacy = await channel.declare_exchange(
-                AbstractPRFeedbackService.EXCHANGE_SCAN_POLLING_LEGACY,
+                PRQueueConstants.EXCHANGE_SCAN_POLLING_LEGACY,
                 aio_pika.ExchangeType.TOPIC,
                 durable=True,
                 internal=True,
@@ -108,13 +109,13 @@ async def setup() -> None:
 
 
             awaited_scans_queue_legacy = await channel.declare_queue(
-                AbstractPRFeedbackService.QUEUE_SCAN_WAIT_LEGACY,
+                PRQueueConstants.QUEUE_SCAN_WAIT_LEGACY,
                 durable=True,
                 arguments={
                     "x-queue-type": "quorum",
                     "x-dead-letter-strategy": "at-least-once",
                     "x-overflow": "reject-publish",
-                    "x-dead-letter-exchange": PRFeedbackService.EXCHANGE_SCAN_POLLING_LEGACY,
+                    "x-dead-letter-exchange": PRQueueConstants.EXCHANGE_SCAN_POLLING_LEGACY,
                 },
             )
             awaited_scans_queue = await channel.declare_queue(
@@ -130,7 +131,7 @@ async def setup() -> None:
 
 
             await awaited_scans_queue_legacy.bind(
-                scan_await_exchange_legacy, PRFeedbackService.ROUTEKEY_POLL_BINDING_LEGACY
+                scan_await_exchange_legacy, PRQueueConstants.ROUTEKEY_POLL_BINDING_LEGACY
             )
             await awaited_scans_queue.bind(
                 scan_await_exchange, ScanPollingService.ROUTEKEY_POLL_BINDING
@@ -139,7 +140,7 @@ async def setup() -> None:
 
 
             polling_scans_queue_legacy = await channel.declare_queue(
-                AbstractPRFeedbackService.QUEUE_SCAN_POLLING_LEGACY,
+                PRQueueConstants.QUEUE_SCAN_POLLING_LEGACY,
                 durable=True,
                 arguments={"x-queue-type": "quorum"},
             )
@@ -152,7 +153,7 @@ async def setup() -> None:
 
             await polling_scans_queue_legacy.bind(
                 polling_delivery_exchange_legacy,
-                AbstractPRFeedbackService.ROUTEKEY_POLL_BINDING_LEGACY,
+                PRQueueConstants.ROUTEKEY_POLL_BINDING_LEGACY,
             )
             await polling_scans_queue.bind(
                 polling_delivery_exchange,
@@ -163,12 +164,12 @@ async def setup() -> None:
 
             # Scan state: Feedback
             pr_feedback_queue = await channel.declare_queue(
-                AbstractPRFeedbackService.QUEUE_FEEDBACK_PR,
+                PRQueueConstants.QUEUE_FEEDBACK_PR,
                 durable=True,
                 arguments={"x-queue-type": "quorum"},
             )
             await pr_feedback_queue.bind(
-                scan_feedback_exchange_pr, PRFeedbackService.ROUTEKEY_FEEDBACK_PR
+                scan_feedback_exchange_pr, PRQueueConstants.ROUTEKEY_FEEDBACK_PR
             )
 
             gen_sarif_queue = await channel.declare_queue(
@@ -182,12 +183,12 @@ async def setup() -> None:
 
             # Scan State: Annotation
             pr_annotate_queue = await channel.declare_queue(
-                AbstractPRFeedbackService.QUEUE_ANNOTATE_PR,
+                PRQueueConstants.QUEUE_ANNOTATE_PR,
                 durable=True,
                 arguments={"x-queue-type": "quorum"},
             )
             await pr_annotate_queue.bind(
-                scan_annotate_exchange_pr, AbstractPRFeedbackService.ROUTEKEY_ANNOTATE_PR
+                scan_annotate_exchange_pr, PRQueueConstants.ROUTEKEY_ANNOTATE_PR
             )
 
         resolver_rmq = await services.resolver.mq_client()
