+++ workflows/messaging/scan_message.py
@@ -1,6 +1,6 @@
 from .base_message import StampedMessage
 from dataclasses import dataclass
-from .. import ScanStates, ScanWorkflow
+from workflows.enums import ScanStates, ScanWorkflow
 
 @dataclass(frozen=True)
 class ScanHeader(StampedMessage):
