+++ workflows/pull_request.py
@@ -1,7 +1,7 @@
 import aio_pika
 from datetime import timedelta
-from workflows.feedback_services import PRFeedbackService
-from workflows import ScanWorkflow, ScanStates, ResultSeverity, ResultStates
+from workflows.feedback_services.pr import PRThreadFeedbackService
+from workflows.enums import ScanWorkflow, ScanStates, ResultSeverity, ResultStates
 from workflows.feedback_workflow_base import AbstractPRFeedbackWorkflow
 from workflows.messaging import ScanAwaitMessage, ScanFeedbackMessage, ScanAnnotationMessage
 from workflows.messaging.util import compute_drop_by_timestamp
@@ -78,7 +78,7 @@ class PullRequestWorkflow(AbstractPRFeedbackWorkflow):
         )
 
     async def workflow_start(self, mq_client : aio_pika.abc.AbstractRobustConnection, moniker : str, projectid : str, scanid : str, **kwargs):
-        topic = PRFeedbackService.make_topic(ScanStates.AWAIT, ScanWorkflow.PR, moniker)
+        topic = PRThreadFeedbackService.make_topic(ScanStates.AWAIT, ScanWorkflow.PR, moniker)
         await self._publish(mq_client, topic, self.__await_msg_factory(projectid, scanid, moniker, **kwargs), 
                             f"{topic} for scan id {scanid} on service {moniker}", CxOneFlowAbstractWorkflowService.EXCHANGE_SCAN_INPUT)
 
@@ -90,11 +90,11 @@ class PullRequestWorkflow(AbstractPRFeedbackWorkflow):
         return self.__enabled
 
     async def feedback_start(self, mq_client : aio_pika.abc.AbstractRobustConnection, moniker : str, projectid : str, scanid : str, **kwargs):
-        topic = PRFeedbackService.make_topic(ScanStates.FEEDBACK, ScanWorkflow.PR, moniker)
+        topic = PRThreadFeedbackService.make_topic(ScanStates.FEEDBACK, ScanWorkflow.PR, moniker)
         await self._publish(mq_client, topic, self.__feedback_msg_factory(projectid, scanid, moniker, **kwargs), 
                             f"{topic} for scan id {scanid} on service {moniker}", CxOneFlowAbstractWorkflowService.EXCHANGE_SCAN_INPUT)
 
     async def annotation_start(self, mq_client : aio_pika.abc.AbstractRobustConnection, moniker : str, projectid : str, scanid : str, annotation : str, **kwargs):
-        topic = PRFeedbackService.make_topic(ScanStates.ANNOTATE, ScanWorkflow.PR, moniker)
+        topic = PRThreadFeedbackService.make_topic(ScanStates.ANNOTATE, ScanWorkflow.PR, moniker)
         await self._publish(mq_client, topic, self.__annotation_msg_factory(projectid, scanid, moniker, annotation, **kwargs), 
                             f"{topic} for scan id {scanid} on service {moniker}", CxOneFlowAbstractWorkflowService.EXCHANGE_SCAN_INPUT)
